Make value sanitization more robust, fix namespace vs. rules precedence, avoid leaking of custom attributes between containers, rename custom.go->annotations.go, hard set workload info annotations, do not gate setting of container attributes in reinvoke, make dependency of meta mutation on oa activation more explicit, avoid key= in resource attribute env var, clean context from ReinvocationRequest, PropagationTest is a helper again.
