+++ pkg/webhook/mutation/pod/mutator/otlp/resourceattributes/attributes_test.go
@@ -8,6 +8,78 @@ import (
 	corev1 "k8s.io/api/core/v1"
 )
 
+func TestSanitizeValue(t *testing.T) {
+	t.Run("empty string is unchanged", func(t *testing.T) {
+		assert.Empty(t, sanitizeValue(""))
+	})
+
+	t.Run("plain alphanum string is unchanged", func(t *testing.T) {
+		assert.Equal(t, "hello", sanitizeValue("hello"))
+	})
+
+	t.Run("env var reference is passed through unchanged", func(t *testing.T) {
+		assert.Equal(t, "$(K8S_POD_NAME)", sanitizeValue("$(K8S_POD_NAME)"))
+	})
+
+	t.Run("spaces are encoded as +", func(t *testing.T) {
+		assert.Equal(t, "hello+world", sanitizeValue("hello world"))
+	})
+
+	t.Run("comma is encoded (would break key=value,key=value format)", func(t *testing.T) {
+		assert.Equal(t, "a%2Cb", sanitizeValue("a,b"))
+	})
+
+	t.Run("equals sign is encoded (would break key=value format)", func(t *testing.T) {
+		assert.Equal(t, "a%3Db", sanitizeValue("a=b"))
+	})
+
+	t.Run("already query-encoded value is idempotent", func(t *testing.T) {
+		// value that went through a previous sanitizeValue call must not be double-encoded
+		once := sanitizeValue("hello world / München")
+		twice := sanitizeValue(once)
+		assert.Equal(t, once, twice)
+	})
+
+	t.Run("value with %XX encoding is normalised and idempotent", func(t *testing.T) {
+		// %20 is a valid percent-encoded space; it gets normalised to + (QueryEscape form)
+		// and must not become %2520 on a second pass
+		normalised := sanitizeValue("hello%20world")
+		assert.Equal(t, "hello+world", normalised)
+		assert.Equal(t, normalised, sanitizeValue(normalised))
+	})
+
+	t.Run("percent sign with invalid sequence is encoded, not double-encoded", func(t *testing.T) {
+		// "100%" contains an incomplete percent sequence; url.QueryUnescape returns an error
+		// so we fall back to encoding the raw string: % becomes %25
+		assert.Equal(t, "100%25", sanitizeValue("100%"))
+	})
+
+	t.Run("percent sign mid-string with invalid sequence is encoded", func(t *testing.T) {
+		// "50%+off" - the % starts an invalid sequence (%+o is not a valid hex pair)
+		assert.Equal(t, "50%25%2Boff", sanitizeValue("50%+off"))
+	})
+
+	// Known caveat: a literal + is indistinguishable from a query-encoded space.
+	// url.QueryUnescape decodes + as space, so it gets re-encoded as + unchanged.
+	// This is acceptable because the OTEL SDK treats both + and %20 as a space anyway.
+	t.Run("literal plus sign is treated as an already-encoded space (known caveat)", func(t *testing.T) {
+		assert.Equal(t, "a+b", sanitizeValue("a+b"))
+	})
+
+	t.Run("already percent-encoded percent sign is idempotent", func(t *testing.T) {
+		// "%25" decodes to "%" which re-encodes to "%25" — stable fixed point
+		assert.Equal(t, "%25", sanitizeValue("%25"))
+		assert.Equal(t, "%25", sanitizeValue(sanitizeValue("%25")))
+	})
+
+	t.Run("unicode characters are percent-encoded", func(t *testing.T) {
+		encoded := sanitizeValue("München")
+		assert.NotEqual(t, "München", encoded)
+		// must be idempotent
+		assert.Equal(t, encoded, sanitizeValue(encoded))
+	})
+}
+
 func TestNewAttributesFromEnv(t *testing.T) {
 	const envName = "OTEL_RESOURCE_ATTRIBUTES"
